+++ locals.tf
@@ -30,6 +30,7 @@ locals {
     name                      = "ip_restriction_cidr_${join("", [1, index(var.authorized_ips, cidr)])}"
     ip_address                = cidr
     virtual_network_subnet_id = null
+    service_tag               = null
     subnet_id                 = null
     priority                  = join("", [1, index(var.authorized_ips, cidr)])
     action                    = "Allow"
@@ -39,10 +40,22 @@ locals {
     name                      = "ip_restriction_subnet_${join("", [1, index(var.authorized_subnet_ids, subnet)])}"
     ip_address                = null
     virtual_network_subnet_id = subnet
+    service_tag               = null
     subnet_id                 = subnet
     priority                  = join("", [1, index(var.authorized_subnet_ids, subnet)])
     action                    = "Allow"
   }]
+
+  service_tags = [for service_tag in var.authorized_service_tags : {
+    name                      = "service_tag_restriction_${join("", [1, index(var.authorized_service_tags, service_tag)])}"
+    ip_address                = null
+    virtual_network_subnet_id = null
+    service_tag               = service_tag
+    subnet_id                 = null
+    priority                  = join("", [1, index(var.authorized_service_tags, service_tag)])
+    action                    = "Allow"
+  }]
+
 
   scm_cidrs = [for cidr in var.scm_authorized_ips : {
     name                      = "scm_ip_restriction_cidr_${join("", [1, index(var.scm_authorized_ips, cidr)])}"
@@ -59,6 +72,16 @@ locals {
     virtual_network_subnet_id = subnet
     subnet_id                 = subnet
     priority                  = join("", [1, index(var.scm_authorized_subnet_ids, subnet)])
+    action                    = "Allow"
+  }]
+
+  scm_service_tags = [for service_tag in var.scm_authorized_service_tags : {
+    name                      = "scm_service_tag_restriction_${join("", [1, index(var.scm_authorized_service_tags, service_tag)])}"
+    ip_address                = null
+    virtual_network_subnet_id = null
+    service_tag               = service_tag
+    subnet_id                 = null
+    priority                  = join("", [1, index(var.scm_authorized_service_tags, service_tag)])
     action                    = "Allow"
   }]