+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changed
 
 Added
   * AZ-331: VNet integration
+  * AZ-424: Add Service Tags (AzureRM 2.42+)
 
 # v3.2.1/v4.0.0 - 2020-11-19